backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.2
Flask-SocketIO==5.5.1
greenlet==3.3.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-engineio==4.12.3
python-socketio==5.15.0
simple-websocket==1.1.0
Werkzeug==3.1.4
wsproto==1.3.2
deepgram-sdk>=3.0.0
python-dotenv>=1.0.0
pymongo>=4.0.0
flask-cors>=4.0.0
